flask/requirements.txt
flask==3.0.3
Flask-Cors==4.0.1
pymongo==4.7.1
Flask-JWT-Extended==4.6.0
Jinja2==3.1.2
python-dotenv==1.0.1
pytz==2023.3.post1
requests==2.31.0